End To End Q&A Chatbot_using_Ollam/requirements.txt
langchain-ollama
langchain_core
python-dotenv
streamlit
langchain_community
python-lsp-server   # ✅ real package, works!
fastapi
uvicorn
bs4
pypdf
fiass-cpu
wikipedia
arxiv